+++ packages/interface/src/App.tsx
@@ -1,25 +1,34 @@
 import { Box } from "@mui/material";
+import { Theme, ThemeProvider } from "@mui/material/styles";
+
 import { TabsMenu } from "./components/layout/TabsMenu";
 import { ValculatorSections } from "./components/layout/Sections";
 
 import { ValculatorContextProvider } from "@valculator/context";
+import { theme as defaultTheme } from "@valculator/theme";
 
-export const Valculator = () => {
+export const Valculator = ({
+  theme,
+}: {
+  theme?: Partial<Theme> | ((outerTheme: Theme) => Theme);
+}) => {
   return (
-    <ValculatorContextProvider>
-      <>
-        <Box
-          padding={2}
-          display={"flex"}
-          flexDirection={"row"}
-          justifyContent={"center"}
-          alignItems={"center"}
-        >
-          <TabsMenu />
-        </Box>
-        <ValculatorSections />
-      </>
-    </ValculatorContextProvider>
+    <ThemeProvider theme={theme ?? defaultTheme}>
+      <ValculatorContextProvider>
+        <>
+          <Box
+            padding={2}
+            display={"flex"}
+            flexDirection={"row"}
+            justifyContent={"center"}
+            alignItems={"center"}
+          >
+            <TabsMenu />
+          </Box>
+          <ValculatorSections />
+        </>
+      </ValculatorContextProvider>
+    </ThemeProvider>
   );
 };