+++ packages/interface/src/main.ts
@@ -1,1 +1,2 @@
 export { Valculator } from "./App";
+export { ShareListButton } from "./components/shared/ShareListButton";