+++ packages/images/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/charlotte-hues/valculator/compare/@valculator/images@0.3.0...@valculator/images@0.3.1) (2024-04-01)
+
+
+### Bug Fixes
+
+* update data -  furniture 2024-04-01 ([#32](https://github.com/charlotte-hues/valculator/issues/32)) ([8fc608a](https://github.com/charlotte-hues/valculator/commit/8fc608addcccd97990c938f18d85814e45821b0b))
+
+
+
+
+
+# [0.3.0](https://github.com/charlotte-hues/valculator/compare/@valculator/images@0.2.0...@valculator/images@0.3.0) (2024-04-01)
+
+
+### Features
+
+* data update 2024-04-01 ([#30](https://github.com/charlotte-hues/valculator/issues/30)) ([e642cc8](https://github.com/charlotte-hues/valculator/commit/e642cc80fbe89304ec8c8f28aaa5e5f1f430c20a))
+
+
+
+
+
 # 0.2.0 (2024-02-25)